+++ software/python/intro comunicacion/instrumental.py
