+++ software/python/practicas/fotoelectrico.py
